requirements.txt
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
pdfplumber==0.10.3
